Add if statement in list to determine how to display notes.

+++ Notes/Home.swift
@@ -26,16 +26,28 @@ struct Home: View {
         
         NavigationView {
             List(self.notes) { note in
-                VStack {
-//                    Image(systemName: "globe")
-//                        .imageScale(.large)
-//                        .foregroundColor(.accentColor)
-                    Text(note.note)
+                
+                if (self.isEditMode == .inactive) {
+                    VStack {
+    //                    Image(systemName: "globe")
+    //                        .imageScale(.large)
+    //                        .foregroundColor(.accentColor)
+                        Text(note.note)
+                    }
+                    .padding()
+                    .onLongPressGesture {
+                        self.showAlert.toggle()
+                        deleteItem = note
+                    }
                 }
-                .padding()
-                .onLongPressGesture {
-                    self.showAlert.toggle()
-                    deleteItem = note
+                else {
+                    HStack {
+                        Image(systemName: "pencil.circle.fill")
+                            .foregroundColor(.yellow)
+                        
+                        Text(note.note)
+                            .padding()
+                    }
                 }
             }
             .alert(isPresented: $showAlert, content: {
